Move all version checking out of mediaChunking and into media
